aps/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
redis==5.2.1
sniffio==1.3.1
starlette==0.44.0
typing_extensions==4.13.2
uvicorn==0.33.0
